move testing constructors
