+++ internal/decision/blockstoremanager.go
@@ -41,16 +41,6 @@ func newBlockstoreManager(
 	}
 }
 
-func newBlockstoreManagerForTesting(
-	ctx context.Context,
-	bs bstore.Blockstore,
-	workerCount int,
-) *blockstoreManager {
-	testPendingBlocksGauge := metrics.NewCtx(ctx, "pending_block_tasks", "Total number of pending blockstore tasks").Gauge()
-	testActiveBlocksGauge := metrics.NewCtx(ctx, "active_block_tasks", "Total number of active blockstore tasks").Gauge()
-	return newBlockstoreManager(ctx, bs, workerCount, testPendingBlocksGauge, testActiveBlocksGauge)
-}
-
 func (bsm *blockstoreManager) start(px process.Process) {
 	px.AddChild(bsm.px)
 	// Start up workers
