+++ internal/decision/engine.go
@@ -210,37 +210,6 @@ func NewEngine(
 	)
 }
 
-func newEngineForTesting(
-	ctx context.Context,
-	bs bstore.Blockstore,
-	bstoreWorkerCount,
-	engineTaskWorkerCount, maxOutstandingBytesPerPeer int,
-	peerTagger PeerTagger,
-	self peer.ID,
-	maxReplaceSize int,
-	scoreLedger ScoreLedger,
-) *Engine {
-	testPendingEngineGauge := metrics.NewCtx(ctx, "pending_tasks", "Total number of pending tasks").Gauge()
-	testActiveEngineGauge := metrics.NewCtx(ctx, "active_tasks", "Total number of active tasks").Gauge()
-	testPendingBlocksGauge := metrics.NewCtx(ctx, "pending_block_tasks", "Total number of pending blockstore tasks").Gauge()
-	testActiveBlocksGauge := metrics.NewCtx(ctx, "active_block_tasks", "Total number of active blockstore tasks").Gauge()
-	return newEngine(
-		ctx,
-		bs,
-		bstoreWorkerCount,
-		engineTaskWorkerCount,
-		maxOutstandingBytesPerPeer,
-		peerTagger,
-		self,
-		maxReplaceSize,
-		scoreLedger,
-		testPendingEngineGauge,
-		testActiveEngineGauge,
-		testPendingBlocksGauge,
-		testActiveBlocksGauge,
-	)
-}
-
 func newEngine(
 	ctx context.Context,
 	bs bstore.Blockstore,
